+++ examples/full-api/src/routa.ts
@@ -1,4 +1,4 @@
-import { createRouta } from "@routa/core";
+import { createRouta } from "@routa-ts/core";
 
 export default createRouta({
 	port: 3000,